move implements_() to ClassCreator and add extends_() to InterfaceCreator
This makes the `InterfaceCreator` usage closer to Java language syntax,
where making an interface extend another interface uses the `extends`
keyword.
